Update redirect-incoming-requests.md
You can't use the second snippet to standardize on a bare domain without a CNAME flattening service.  So, we need to add the steps to set that up.

www and non-www example

add link to doc

Adding link to Pantheon Cloudflare doc

add clarification on SSL

If a person has an SSL cert, they don't need the Cloudflare setup.  Adding that disclaimer.

reformat the Note

made the note look like others on the page

spacing